+++ src/Encamina.Enmarcha.SemanticKernel/MemoryManager.cs
@@ -20,9 +20,9 @@ public class MemoryManager : IMemoryManager
 {
     private const string ChunkSize = @"chunkSize";
 
-    private readonly Kernel kernel;
-
     private readonly ILogger logger;
+
+    private readonly ITextEmbeddingGenerationService textEmbeddingGenerationService;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="MemoryManager"/> class.
@@ -31,18 +31,19 @@ public class MemoryManager : IMemoryManager
     /// A valid instance of <see cref="Kernel"/>, used to get the configured text embeddings generation service (<see cref="ITextEmbeddingGenerationService"/>) required by this manager.
     /// </param>
     /// <param name="memoryStore">A valid instance of a <see cref="IMemoryStore"/> to manage.</param>
-    public MemoryManager(Kernel kernel, IMemoryStore memoryStore)
+    public MemoryManager(IMemoryStore memoryStore, ITextEmbeddingGenerationService textEmbeddingGenerationService, ILogger<MemoryManager> logger)
     {
-        this.kernel = kernel;
-        logger = kernel.LoggerFactory.CreateLogger<MemoryManager>();
+        this.logger = logger;
         MemoryStore = memoryStore;
+
+        this.textEmbeddingGenerationService = textEmbeddingGenerationService;
     }
 
     /// <inheritdoc/>
     public IMemoryStore MemoryStore { get; init; }
 
     /// <inheritdoc/>
-    public virtual async Task UpsertMemoryAsync(string memoryId, string collectionName, IEnumerable<string> chunks, CancellationToken cancellationToken, IDictionary<string, string> metadata = null)
+    public virtual async Task UpsertMemoryAsync(string memoryId, string collectionName, IEnumerable<string> chunks, IDictionary<string, string> metadata = null, Kernel kernel = null, CancellationToken cancellationToken = default)
     {
         var memoryChunkSize = await GetChunkSize(memoryId, collectionName, cancellationToken);
 
@@ -51,7 +52,7 @@ public class MemoryManager : IMemoryManager
             await DeleteMemoryAsync(memoryId, collectionName, memoryChunkSize, cancellationToken);
         }
 
-        await SaveChunks(memoryId, collectionName, chunks, metadata, cancellationToken);
+        await SaveChunks(memoryId, collectionName, chunks, metadata, kernel, cancellationToken);
     }
 
     /// <inheritdoc/>
@@ -86,7 +87,7 @@ public class MemoryManager : IMemoryManager
     }
 
     /// <inheritdoc/>
-    public virtual async IAsyncEnumerable<string> BatchUpsertMemoriesAsync(string collectionName, IDictionary<string, MemoryContent> memoryContents, [EnumeratorCancellation] CancellationToken cancellationToken)
+    public virtual async IAsyncEnumerable<string> BatchUpsertMemoriesAsync(string collectionName, IDictionary<string, MemoryContent> memoryContents, Kernel kernel = null, [EnumeratorCancellation] CancellationToken cancellationToken = default)
     {
         var memoryRecords = new Collection<MemoryRecord>();
 
@@ -103,7 +104,7 @@ public class MemoryManager : IMemoryManager
                 for (var i = 0; i < totalChunks; i++)
                 {
                     var chunk = memoryContent.Chunks.ElementAt(i);
-                    var embedding = await kernel.GetRequiredService<ITextEmbeddingGenerationService>().GenerateEmbeddingAsync(chunk, kernel, cancellationToken);
+                    var embedding = await textEmbeddingGenerationService.GenerateEmbeddingAsync(chunk, kernel, cancellationToken);
                     memoryRecords.Add(MemoryRecord.LocalRecord($@"{memoryContentId}-{i}", chunk, null, embedding, JsonSerializer.Serialize(memoryContent.Metadata)));
                 }
             }
@@ -139,7 +140,7 @@ public class MemoryManager : IMemoryManager
         await MemoryStore.RemoveBatchAsync(collectionName, Enumerable.Range(0, chunkSize).Select(i => BuildMemoryIdentifier(memoryId, i)), cancellationToken);
     }
 
-    private async Task SaveChunks(string memoryid, string collectionName, IEnumerable<string> chunks, IDictionary<string, string> metadata, CancellationToken cancellationToken)
+    private async Task SaveChunks(string memoryid, string collectionName, IEnumerable<string> chunks, IDictionary<string, string> metadata, Kernel kernel = null, CancellationToken cancellationToken = default)
     {
         metadata ??= new Dictionary<string, string>();
 
@@ -154,7 +155,7 @@ public class MemoryManager : IMemoryManager
         for (var i = 0; i < chunksCount; i++)
         {
             var chunk = chunks.ElementAt(i);
-            var embedding = await kernel.GetRequiredService<ITextEmbeddingGenerationService>().GenerateEmbeddingAsync(chunk, kernel, cancellationToken);
+            var embedding = await textEmbeddingGenerationService.GenerateEmbeddingAsync(chunk, kernel, cancellationToken);
             memoryRecords.Add(MemoryRecord.LocalRecord(BuildMemoryIdentifier(memoryid, i), chunk, null, embedding, metadataJson));
         }